Added Point Density Hexagon Layer (#370)
Changes included
PointDensityHexagonLayer

PointDensityHexagonLayer is a layer wrapper that aggregates points into hexagonal bins
It takes an array of points and the radius of hexagon bin in meters
Color and height of each hexagon are scaled by the density of the bins, that is, the number of points it contains.
By default PointDensityHexagonLayer uses d3-hexbin as hexagon aggregator it also takes hexagonAggregator as a prop
HexagonLayer

HexagonLayer now also takes in radius in meter and angle in radian as props to calculate primitive hexagon. If radius and angle are provided, hexagonVertices will be ignored.
Change radiusScale to coverage for clarity. coverage will be clamped between 0 - 1
Added elevationScale to scale elevation.
GridLayer

Added elevationScale to scale elevation.
LayerControl

Added ColorRangePicker