Workaround for invalid attribute index access
In an application with heavy path layer usage the typed array manager
can get to a point where the number of points exactly fits in the
allocated buffer.  This problem is this does not appear to be accounting
for the shaderAttributes and results the runtime error:

  'glDrawElementsInstancedANGLE: attempt to access out of range vertices in attribute'

This extra padding 3 * 6 is based on how the type array manager needs to allocate
for largest case in fp64.