Enable custom_map_style and file encoding for HTML on Windows
Add parameter to Deck class that allows passing in a dictionary definition for a Mapbox style specification. This allows a user to define custom styles, without needing to host a style.json, for use within disconnected environments, or with self-hosted WMS or XYZ tile sets. Additionally, fix character set encoding on Windows environments when exporting a Deck visualization to HTML.