Migrate jupyter-widget to a jupyterlab 3 prebuilt extension
The main thing is modifying package.json and webpack.config.js to
work with the jupyterlab builder. Since the previous build was
just using webpack, we let the builder integrate the webpack config
from that build with the labextension webpack. Required some changes
to remove the original dist directory so everything just goes to a
pydeck/labextension directory now.