Make canvas readonly instead of protected
Change typescript annotation to make the Deck instance canvas readonly instead of protected. Right now there isn't a way to access the canvas instance [as seems to be suggested](https://github.com/visgl/deck.gl/issues/4436#issuecomment-610472868), as Typescript forbids it. It's possible to cast the deck instance as `any` before getting the canvas prop but this way provides a formalized accessor.