doc: using-with-mapbox: fix broken link
On the using-with-mapbox page of the docs, there is a broken link to the
`react-map-gl` docs under the Compatibility with Mapbox GL JS forks
section. If you click "these instructions" at the end of the section,
you'll navigate to:
http://visgl.github.io/react-map-gl/docs/get-started/get-started#using-with-a-mapbox-gl-fork,
which ends up showing a Page Not Found error.

Let's update the documentation such that folks reading it will be linked
to the right place for finding more info.